requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
celery>=5.2.0
redis>=4.0.0
django-celery-results>=2.2.0
requests>=2.25.1,<3.0.0
django-filter>=21.1
gunicorn==20.1.0
whitenoise==6.5.0